refactor: split spez into different files
